Implement basic-authentication for InfluxDB requests
The Kamon-InfluxDB extension that worked with kamon 0.6.x allowed for specifying
username + password for InfluxDB. This functionality has been lost with the
current implementation. The previous implementation was insecure passing the
password as a query parameter in the URL. Using basic authentication, which
the OkHttpClient supports, we have a more secure way of passing this information.