add support for the DuckDB and AWS Advanced Wrapper drivers in JDBC (#1401)
Adds support for the two drivers and includes a few additional tests for
the wrapper, both with and withouth the Hikari pool.

Also during testing I realized that we have to explicitly ignore some of
the AWS wrapper classes to avoid having double spans for each database
interaction. Funny enough, in that setup we have Hikari wrapping the AWS
wrapper which wraps the actual wrapper